Fixing BigDecimal always mapped to Long
When trying to save values of type BigDecimal we encountered loss of precision. We found out that if you try to save values of type BigDecimal micronaut-data will fall-through to the LONG block of the Switch Statement and convert any BigDecimal to Long values without any complaint. You just see you lost your decimal digits in the database.