Fix NullPointerException for Embedded null values
If we have Embedded values in an entity and we have an instance of an entity where the embedded value is null, we get a NullPointerException.

i.e. we have a Class `Totals` which contains several subProperties of Type `@Embedded Monitary` (costs, turnover). Which each have the properties `currency` and `amount`.

Now if we have an instance where only `Monitary costs` is filled but `Monitary turnover` is null we get a NullpointerException at L150 because `value` in L149 is null.