Use new Micronaut Test Resources features (#2171)
This commit introduces the use of the test resources scope. This
simplifies the test setup by avoiding the direct calls to the
test resources client and supporting a more declarative style.

There is one catch, which is that some tests do not use the
`@MicronautTest` annotation, so they cannot benefit from the
`TestPropertyProviderFactory`, which means that these tests
have to implement the service loading logic themselves.

Hopefully there are not so many occurrences of this.